Remove getSanitizedData function.
It hides the fact that there is a (quite complex) getData call behind it,
so we have code like this:

    pool[emojiId] = getData(emoji, null, null, this.data)
    this.emojis[emojiId] = getSanitizedData(emoji, null, null, this.data)

It is not obvious, but we actually call getData twice here.
Now this is changed to

    pool[emojiId] = getData(emoji, null, null, this.data)
    this.emojis[emojiId] = getSanitizedData(pool[emojiId])

And, since, there is no getSanitizedData anymore, we guarantee that no
such error will be made in the future.